color sharps/flats differently
