+++ src/components/PanCello/CNotes.jsx
@@ -21,6 +21,17 @@ class CNotes extends Component {
     return (
       <>
         {/* C NOTES */}
+        <g id="C3">
+          <ellipse
+            style={{ fill: this.props.colors.C3, stroke: 'black' }}
+            cx="733"
+            cy="604.5"
+            rx="159.5"
+            ry="103"
+            onTouchStart={this.props.isTouch ? () => this.playNote('C3') : null}
+            onClick={this.props.isTouch ? null : () => this.playNote('C3')}
+          />
+        </g>
         <g id="C4">
           <ellipse
             transform="matrix(0.8284 -0.5601 0.5601 0.8284 -179.2137 370.7318)"
@@ -33,40 +44,6 @@ class CNotes extends Component {
             onClick={this.props.isTouch ? null : () => this.playNote('C4')}
           />
         </g>
-        <g id="CSh4">
-          <ellipse
-            transform="matrix(0.8284 -0.5601 0.5601 0.8284 -469.2018 688.0895)"
-            style={{ fill: this.props.colors.CSh4, stroke: 'black' }}
-            cx="888.505"
-            cy="1109.88"
-            rx="99.607"
-            ry="77.195"
-            onTouchStart={this.props.isTouch ? () => this.playNote('CSh4') : null}
-            onClick={this.props.isTouch ? null : () => this.playNote('CSh4')}
-          />
-        </g>
-        <g id="C3">
-          <ellipse
-            style={{ fill: this.props.colors.C3, stroke: 'black' }}
-            cx="733"
-            cy="604.5"
-            rx="159.5"
-            ry="103"
-            onTouchStart={this.props.isTouch ? () => this.playNote('C3') : null}
-            onClick={this.props.isTouch ? null : () => this.playNote('C3')}
-          />
-        </g>
-        <g id="CSh3">
-          <ellipse
-            style={{ fill: this.props.colors.CSh3, stroke: 'black' }}
-            cx="1106"
-            cy="1235.5"
-            rx="159.5"
-            ry="103"
-            onTouchStart={this.props.isTouch ? () => this.playNote('CSh3') : null}
-            onClick={this.props.isTouch ? null : () => this.playNote('CSh3')}
-          />
-        </g>
         <g id="C5">
           <ellipse
             transform="matrix(0.6259 -0.7799 0.7799 0.6259 -71.1086 695.6706)"
@@ -77,18 +54,6 @@ class CNotes extends Component {
             ry="38.5"
             onTouchStart={this.props.isTouch ? () => this.playNote('C5') : null}
             onClick={this.props.isTouch ? null : () => this.playNote('C5')}
-          />
-        </g>
-        <g id="CSh5">
-          <ellipse
-            transform="matrix(0.6259 -0.7799 0.7799 0.6259 -422.9115 1222.2793)"
-            style={{ fill: this.props.colors.CSh5, stroke: 'black' }}
-            cx="1062.566"
-            cy="1051.954"
-            rx="55.601"
-            ry="38.5"
-            onTouchStart={this.props.isTouch ? () => this.playNote('CSh5') : null}
-            onClick={this.props.isTouch ? null : () => this.playNote('CSh5')}
           />
         </g>
       </>
