+++ src/components/PanCello/GNotes.jsx
@@ -33,18 +33,6 @@ class GNotes extends Component {
             onClick={this.props.isTouch ? null : () => this.playNote('G3')}
           />
         </g>
-        <g id="GSh3">
-          <ellipse
-            transform="matrix(0.3602 -0.9329 0.9329 0.3602 -615.655 1254.9376)"
-            style={{ fill: this.props.colors.GSh3, stroke: 'black' }}
-            cx="607.016"
-            cy="1076.278"
-            rx="125.535"
-            ry="87.343"
-            onTouchStart={this.props.isTouch ? () => this.playNote('GSh3') : null}
-            onClick={this.props.isTouch ? null : () => this.playNote('GSh3')}
-          />
-        </g>
         <g id="G4">
           <ellipse
             transform="matrix(0.9883 -0.1526 0.1526 0.9883 -135.018 194.1195)"
@@ -55,18 +43,6 @@ class GNotes extends Component {
             ry="51.345"
             onTouchStart={this.props.isTouch ? () => this.playNote('G4') : null}
             onClick={this.props.isTouch ? null : () => this.playNote('G4')}
-          />
-        </g>
-        <g id="GSh4">
-          <ellipse
-            transform="matrix(0.9883 -0.1526 0.1526 0.9883 -143.7587 80.2567)"
-            style={{ fill: this.props.colors.GSh4, stroke: 'black' }}
-            cx="450.861"
-            cy="976.481"
-            rx="74.882"
-            ry="51.345"
-            onTouchStart={this.props.isTouch ? () => this.playNote('GSh4') : null}
-            onClick={this.props.isTouch ? null : () => this.playNote('GSh4')}
           />
         </g>
       </>
