+++ src/components/PanCello/PanCello.jsx
@@ -5,11 +5,16 @@ import './PanCello.css';
 // svg imports
 import ANotes from './ANotes';
 import BNotes from './BNotes';
+import BbNotes from './BbNotes';
 import CNotes from './CNotes';
+import CShNotes from './CShNotes';
 import DNotes from './DNotes';
 import ENotes from './ENotes';
+import EbNotes from './EbNotes';
 import FNotes from './FNotes';
+import FShNotes from './FShNotes';
 import GNotes from './GNotes';
+import GShNotes from './GShNotes';
 
 class PanCello extends Component {
   state = {
@@ -59,11 +64,16 @@ class PanCello extends Component {
 
           <ANotes />
           <BNotes />
+          <BbNotes />
           <CNotes />
+          <CShNotes />
           <DNotes />
           <ENotes />
+          <EbNotes />
           <FNotes />
+          <FShNotes />
           <GNotes />
+          <GShNotes />
         </svg>
       </div>
     );
