+++ src/components/PanTenor/BNotes.jsx
@@ -3,7 +3,7 @@ import NOTES from '../../constants/tenorNotes';
 import COLORS from '../../constants/colors';
 import { connect } from 'react-redux';
 
-class BNotes extends Component {
+class BbNotes extends Component {
 
   playNote = (note) => {
     NOTES[note].play();
@@ -11,8 +11,8 @@ class BNotes extends Component {
       type: 'PLAY_TENOR_NOTE',
       payload: {
         note: note,
-        color: this.props.displayColors ? COLORS.pink : COLORS.colorless,
-        highlight: this.props.displayColors ? COLORS.pinkHighlight : COLORS.colorlessHighlight
+        color: this.props.displayColors ? COLORS.violet : COLORS.colorless,
+        highlight: this.props.displayColors ? COLORS.violetHighlight : COLORS.colorlessHighlight
       }
     })
   }
@@ -33,18 +33,6 @@ class BNotes extends Component {
             onClick={this.props.isTouch ? null : () => this.playNote('B4')}
           />
         </g>
-        <g id="Bb4">
-          <ellipse
-            transform="matrix(0.9599 -0.2802 0.2802 0.9599 -247.888 89.9579)"
-            style={{ fill: this.props.colors.Bb4, stroke: 'black' }}
-            cx="190.625"
-            cy="911.805"
-            rx="129.46"
-            ry="101.873"
-            onTouchStart={this.props.isTouch ? () => this.playNote('Bb4') : null}
-            onClick={this.props.isTouch ? null : () => this.playNote('Bb4')}
-          />
-        </g>
         <g id="B5">
           <circle
             style={{ fill: this.props.colors.B5, stroke: 'black' }}
@@ -53,16 +41,6 @@ class BNotes extends Component {
             r="57.5"
             onTouchStart={this.props.isTouch ? () => this.playNote('B5') : null}
             onClick={this.props.isTouch ? null : () => this.playNote('B5')}
-          />
-        </g>
-        <g id="Bb5">
-          <circle
-            style={{ fill: this.props.colors.Bb5, stroke: 'black' }}
-            cx="468"
-            cy="931"
-            r="61.5"
-            onTouchStart={this.props.isTouch ? () => this.playNote('Bb5') : null}
-            onClick={this.props.isTouch ? null : () => this.playNote('Bb5')}
           />
         </g>
       </>
@@ -76,4 +54,4 @@ const mapStateToProps = state => ({
   isTouch: state.isTouch,
 });
 
-export default connect(mapStateToProps)(BNotes);
+export default connect(mapStateToProps)(BbNotes);
