+++ src/constants/colors.js
@@ -3,16 +3,26 @@ const COLORS = {
   blueHighlight: '#00157C',
   pink: '#FFBFCA',
   pinkHighlight: '#FF577B',
+  violet: '#EE82EE',
+  violetHighlight: '#9400D3',
   red: 'red',
-  redHighlight: '#930E00',
+  redHighlight: '#B20000',
+  orangeRed: '#FF4500',
+  orangeRedHighlight: '#CC3700',
   orange: 'orange',
   orangeHighlight: '#EA7A00',
   yellow: 'yellow',
   yellowHighlight: '#D8D800',
+  maroon: '#B22222',
+  maroonHighlight: '#800000',
   green: 'green',
   greenHighlight: '#224900',
+  mint:'#98FF98',
+  mintHighlight: '#A4F3A4',
   indigo: '#A9D7E4',
   indigoHighlight: '#23A9C9',
+  aqua: '#00FFFF',
+  aquaHighlight: '#00CCCC',
   colorless: 'white',
   colorlessHighlight: '#DCDCDC',
 }
