+++ src/redux/reducers/bassReducer.js
@@ -5,21 +5,20 @@ const colors = {
   B2: COLORS.pink,
   Bb2: COLORS.pink,
   C2: COLORS.red,
-  CSh2: COLORS.red,
+  CSh2: COLORS.orangeRed,
   C3: COLORS.red,
-  CSh3: COLORS.red,
+  CSh3: COLORS.orangeRed,
   D2: COLORS.orange,
   D3: COLORS.orange,
   E2: COLORS.yellow,
-  Eb2: COLORS.yellow,
+  Eb2: COLORS.maroon,
   E3: COLORS.yellow,
-  Eb3: COLORS.yellow,
-  FSh2: COLORS.green,
+  Eb3: COLORS.maroon,
   F2: COLORS.green,
-  FSh3: COLORS.green,
+  FSh2: COLORS.mint,
   F3: COLORS.green,
   G2: COLORS.indigo,
-  GSh2: COLORS.indigo,
+  GSh2: COLORS.aqua,
 }
 
 // create noColor object with notes as keys, 'white' as values
