+++ src/redux/reducers/celloReducer.js
@@ -3,31 +3,31 @@ import COLORS from '../../constants/colors';
 const colors = {
   A3: COLORS.blue,
   A4: COLORS.blue,
-  B3: COLORS.pink,
+  B3: COLORS.violet,
   Bb3: COLORS.pink,
-  B4: COLORS.pink,
+  B4: COLORS.violet,
   Bb4: COLORS.pink,
   C3: COLORS.red,
-  CSh3: COLORS.red,
+  CSh3: COLORS.orangeRed,
   C4: COLORS.red,
-  CSh4: COLORS.red,
+  CSh4: COLORS.orangeRed,
   C5: COLORS.red,
-  CSh5: COLORS.red,
+  CSh5: COLORS.orangeRed,
   D3: COLORS.orange,
   D4: COLORS.orange,
   D5: COLORS.orange,
   E3: COLORS.yellow,
-  Eb3: COLORS.yellow,
+  Eb3: COLORS.maroon,
   E4: COLORS.yellow,
-  Eb4: COLORS.yellow,
-  FSh3: COLORS.green,
+  Eb4: COLORS.maroon,
   F3: COLORS.green,
-  FSh4: COLORS.green,
+  FSh3: COLORS.mint,
   F4: COLORS.green,
+  FSh4: COLORS.mint,
   G3: COLORS.indigo,
-  GSh3: COLORS.indigo,
+  GSh3: COLORS.aqua,
   G4: COLORS.indigo,
-  GSh4: COLORS.indigo,
+  GSh4: COLORS.aqua,
 }
 
 // create noColor object with notes as keys, 'white' as values
