+++ src/redux/reducers/secondReducer.js
@@ -4,35 +4,35 @@ const colors = {
   A3: COLORS.blue,
   A4: COLORS.blue,
   A5: COLORS.blue,
-  B3: COLORS.pink,
+  B3: COLORS.violet,
   Bb3: COLORS.pink,
-  B4: COLORS.pink,
+  B4: COLORS.violet,
   Bb4: COLORS.pink,
   Bb5: COLORS.pink,
   C4: COLORS.red,
-  CSh4: COLORS.red,
+  CSh4: COLORS.orangeRed,
   C5: COLORS.red,
-  CSh5: COLORS.red,
+  CSh5: COLORS.orangeRed,
   C6: COLORS.red,
-  CSh6: COLORS.red,
+  CSh6: COLORS.orangeRed,
   D4: COLORS.orange,
   D5: COLORS.orange,
   E4: COLORS.yellow,
-  Eb4: COLORS.yellow,
+  Eb4: COLORS.maroon,
   E5: COLORS.yellow,
-  Eb5: COLORS.yellow,
+  Eb5: COLORS.maroon,
   F3: COLORS.green,
-  FSh3: COLORS.green,
+  FSh3: COLORS.mint,
   F4: COLORS.green,
-  FSh4: COLORS.green,
+  FSh4: COLORS.mint,
   F5: COLORS.green,
-  FSh5: COLORS.green,
+  FSh5: COLORS.mint,
   G3: COLORS.indigo,
-  GSh3: COLORS.indigo,
+  GSh3: COLORS.aqua,
   G4: COLORS.indigo,
-  GSh4: COLORS.indigo,
+  GSh4: COLORS.aqua,
   G5: COLORS.indigo,
-  GSh5: COLORS.indigo,
+  GSh5: COLORS.aqua,
 }
 
 // create noColor object with notes as keys, 'white' as values
