+++ src/redux/reducers/tenorReducer.js
@@ -3,33 +3,33 @@ import COLORS from '../../constants/colors';
 const colors = {
   A4: COLORS.blue,
   A5: COLORS.blue,
-  B4: COLORS.pink,
+  B4: COLORS.violet,
   Bb4: COLORS.pink,
-  B5: COLORS.pink,
+  B5: COLORS.violet,
   Bb5: COLORS.pink,
   C4: COLORS.red,
-  CSh4: COLORS.red,
+  CSh4: COLORS.orangeRed,
   C5: COLORS.red,
-  CSh5: COLORS.red,
+  CSh5: COLORS.orangeRed,
   C6: COLORS.red,
-  CSh6: COLORS.red,
+  CSh6: COLORS.orangeRed,
   D4: COLORS.orange,
   D5: COLORS.orange,
   D6: COLORS.orange,
   E4: COLORS.yellow,
-  Eb4: COLORS.yellow,
+  Eb4: COLORS.maroon,
   E5: COLORS.yellow,
-  Eb5: COLORS.yellow,
+  Eb5: COLORS.maroon,
   E6: COLORS.yellow,
-  Eb6: COLORS.yellow,
-  FSh4: COLORS.green,
+  Eb6: COLORS.maroon,
+  FSh4: COLORS.mint,
   F4: COLORS.green,
-  FSh5: COLORS.green,
+  FSh5: COLORS.mint,
   F5: COLORS.green,
   G4: COLORS.indigo,
-  GSh4: COLORS.indigo,
+  GSh4: COLORS.aqua,
   G5: COLORS.indigo,
-  GSh5: COLORS.indigo,
+  GSh5: COLORS.aqua,
 }
 
 // create noColor object with notes as keys, 'white' as values
